[#201] Brought many of william's requested changes - Refactored capture_image function into another function in the stream_capture.py file - Stream captures now go in a directory with the camera name on it
