+++ robot/basestation/static/js/elements/tools/utils.js
@@ -1,22 +1,3 @@
-// Keep the state of changable text of a button when called on upon page load
-function keepText (button) {
-  if (window.localStorage.getItem(button) == 'text-swap') {
-    toggleText(button)
-  }
-}
-
-// Toggle text of a button
-function toggleText (button) {
-  if (button.text() == button.data('text-swap')) {
-    button.text(button.data('text-original'))
-    window.localStorage.setItem(button, 'text-original')
-  } else {
-    button.data('text-original', button.text())
-    button.text(button.data('text-swap'))
-    window.localStorage.setItem(button, 'text-swap')
-  }
-}
-
 // Applies modulo and if the value is negative, cycle it.
 function negativeModulo (number, modulo) {
   while (number < 0) {
@@ -36,12 +17,4 @@ function rotateElement (element, angle) {
     '-webkit-transform': 'rotate(' + angle + 'deg)',
     '-o-transform': 'rotate(' + angle + 'deg)'
   })
-}
-
-// Generates unique names for html elements with a common class
-function addIdentifiers(cssClass) {
-  $(cssClass).each((i, e) => {
-    let removedPrefix = cssClass.substr(1); 
-    $(e).addClass(removedPrefix + "-" + i);
-  });
 }
